Fix regression in multi-user Audio Record
In the case of audio input, the address specifies the user id.
The client_id has to be further retrieved from the user id.
The previous fix for KW causes the audio input APIs to return
error, since it validates the user id instead of client id.
This fix does the check after get_client_id_from_user_id call.

Change-Id: I614843e4fd3de577e3c93c0a5a2936db182b0594
Tracked-On: OAM-101920
Signed-off-by: Anoob Anto K <anoob.anto.kodankandath@intel.com>