Revert Fix for recording glitch as it causes AV Sync issue
This reverts commit 02ed146fd6f614756563cbf33106b3ed93b94e50.
The above fix was done to prevent glitches in audio, by waiting
additional 2ms for each packet to arrive. But these additional
delays add up and reduces the frame rate of the recorded audio.

Streaming usecases rely on udp protocol, and doesn't guarantee
the rate of packet transfer. Instead, it sends the latest packet
available at a particular time to achieve synchronization.
Since Android framework expects the audio packets in specific
intervals, we have to introduce silent frames for missed packets
to achieve frame-rate and synchronization.

Tracked-On: OAM-100474
Signed-off-by: Anoob Anto K <anoob.anto.kodankandath@intel.com>